add the customer needs and to verify the tests

+++ src/Controller/TaskController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Controller;
+namespace App\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -8,8 +8,12 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 //use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 use Doctrine\ORM\EntityManagerInterface;
-use AppBundle\Entity\Task;
-use AppBundle\Form\TaskType;
+use App\Entity\Task;
+use App\Form\TaskType;
+use Symfony\Component\Notifier\Notification\Notification;
+use Symfony\Component\Notifier\NotifierInterface;
+use App\Entity\User;
+use App\Form\UserType;
 
 
 class TaskController extends AbstractController
@@ -40,6 +44,12 @@ class TaskController extends AbstractController
 
             $task = $form->getData();
 
+            if($this->getUser()){
+                $user = $this->getUser();
+
+                $task->setAuthor($user->getUsername());
+            }
+
             $entityManager->persist($task);
             $entityManager->flush();
 
@@ -67,12 +77,18 @@ class TaskController extends AbstractController
 
         $task = new Task();
 
+        $task->setTitle($task_db->getTitle());
+        $task->setContent($task_db->getContent());
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $task_db = $form->getData();
+            $task = $form->getData();
+
+            $task_db->setTitle($task->getTitle());
+            $task_db->setContent($task->getContent());
 
             $entityManager->persist($task_db);
             $entityManager->flush();
@@ -87,7 +103,7 @@ class TaskController extends AbstractController
 
         return $this->render('task/edit.html.twig', [
             'form' => $form->createView(), 
-            'task' => $task,
+            'task' => $task_db,
         ]);
     }
 
@@ -124,10 +140,28 @@ class TaskController extends AbstractController
             );
         }
 
-        $entityManager->remove($task_db);
-        $entityManager->flush();
+        if($this->getUser()){
+            $user = $this->getUser();
 
-        $notifier->send(new Notification('La tâche a bien été supprimée.', ['browser']));
+            if($user->getUsername()==$task_db->getAuthor() || ($task_db->getAuthor()=="anonyme" && $user->getRole()=="ROLE_ADMIN")){
+                $entityManager->remove($task_db);
+                $entityManager->flush();
+
+                $notifier->send(new Notification('La tâche a bien été supprimée.', ['browser']));
+            }
+
+            /*if($task_db->getAuthor()=="anonyme" && $user-getRole()=="ROLE_ADMIN"){
+                $entityManager->remove($task_db);
+                $entityManager->flush();
+
+                $notifier->send(new Notification('La tâche a bien été supprimée.', ['browser']));
+            }*/
+
+        }
+
+        
+
+        
 
         /*$em = $this->getDoctrine()->getManager();
         $em->remove($task);
